Extract PracticeInitialScreen scaffold + usePracticeDeck; share RadioRow
Both practice InitialScreens' identical structure (heading, JLPT/bookmark
filters, shared session toggles, sticky start footer) and identical deck
data flow now come from shared-practice; only the mode-specific session
options remain in each screen. SpeedKatakana's verbatim local ToggleRow
and RadioRow copies replaced with the shared components.
