+++ src/components/shared-practice/index.ts
@@ -3,9 +3,12 @@ export { BlurredGloss } from "./BlurredGloss";
 export { Stat } from "./CountUpStat";
 export { EndScreenLayout } from "./EndScreenLayout";
 export { EndSession } from "./EndSession";
+export { PracticeInitialScreen } from "./PracticeInitialScreen";
 export { PracticeShell } from "./PracticeShell";
+export { RadioRow } from "./RadioRow";
 export { RecapTile } from "./RecapTile";
 export { ToggleRow } from "./ToggleRow";
+export { usePracticeDeck } from "./use-practice-deck";
 export type {
   DeckFilterSettings,
   PracticeItem,
